Move ml_estimator submodules up a level & remove directories
Now that we save pickled model files in a different directory,
there's no need to have a directory per estimator class.
