+++ src/machine_learning/ml_estimators/__init__.py
@@ -1,6 +1,5 @@
 """ML model classes and associated data classes"""
 
-from .base_ml_estimator import BaseMLEstimator
 from .benchmark_estimator import BenchmarkEstimator
 from .bagging_estimator import BaggingEstimator
 from .stacking_estimator import StackingEstimator
