--- src/machine_learning/ml_estimators/bagging_estimator/__init__.py
@@ -1,1 +1,0 @@
-from .bagging_estimator import BaggingEstimator
