--- src/machine_learning/ml_estimators/benchmark_estimator/__init__.py
@@ -1,1 +1,0 @@
-from .benchmark_estimator import BenchmarkEstimator
