--- src/machine_learning/ml_estimators/stacking_estimator/__init__.py
@@ -1,1 +1,0 @@
-from .stacking_estimator import StackingEstimator
